// src/app/model/user.ts
export class User {

    id: number;
    firstName: string;
    lastName: string;
    loginId: string;
    password: string;
    address: string;
    about: string;
    city: string;
    state: string;
    zip: string;
    phone: string;
    admin: boolean;
    user: boolean;
    userAdmin: boolean;
    audit: boolean;
    readOnly: boolean;
    fine: boolean;
    createdDate: string;
    updateDate: string;
    updateBy: string;

}